Update kbmod_analoysis_demo notebook
